refactor works
